# acpx/docs/schema/acpx-http.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "acpx gateway HTTP/WS transport envelope",
    "version": "0.1.0",
    "description": "Describes the HTTP transport envelope acpx-server wraps its JSON-RPC body in -- the two fixed paths (POST /rpc, GET /ws) and the three header-level contracts (Authorization, X-Acpx-Tenant, X-Acpx-Profile) a JSON Schema/OpenRPC body document has no place to describe. For the JSON-RPC body's own shape (every method's params/result), see docs/schema/acpx.openrpc.json. stdio is a separate transport with no headers and is not described here -- see acpx.openrpc.json's servers list."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8790",
      "description": "default ACPX_HTTP_BIND"
    }
  ],
  "paths": {
    "/rpc": {
      "post": {
        "summary": "JSON-RPC-over-HTTP",
        "description": "Body is a raw JSON-RPC 2.0 request; response body is the JSON-RPC response.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "description": "Bearer token, checked in constant time against ACPX_AUTH_TOKEN when that env var is set on the server (acpx-server/src/transport/http.rs's AuthConfig). Unset ACPX_AUTH_TOKEN means auth is disabled entirely -- this header is then ignored if sent and not required if omitted. When ACPX_AUTH_TOKEN is set, a missing or mismatched header yields 401 Unauthorized with a JSON-RPC-shaped error body (see the shared errorResponse schema on this document's responses).",
            "schema": {
              "type": "string",
              "pattern": "^Bearer .+$"
            },
            "example": "Bearer sk-acpx-..."
          },
          {
            "name": "X-Acpx-Tenant",
            "in": "header",
            "required": false,
            "description": "Self-declared tenant partition key (acpx-tenant-isolation plan) applied to every method on this request/connection, not just session/new. Absent or not valid UTF-8 is treated identically to the implicit default tenant (\"default\") -- this is a data partition, not an auth gate, so it fails open rather than rejecting the request. For WS, this header is read once at upgrade time and cached for the connection's entire lifetime (a WS client is one fixed tenant for its whole connection).",
            "schema": {
              "type": "string",
              "default": "default"
            },
            "example": "acme-corp"
          },
          {
            "name": "X-Acpx-Profile",
            "in": "header",
            "required": false,
            "description": "Explicit profile selection for a session/new request on this transport -- POST /rpc only (no WS equivalent: WS headers are only present at the initial upgrade, not per-frame, so a WS client that wants managed mode must instead set params._acpx.profile inline on its session/new frame). Highest precedence: overwrites any inline params._acpx.profile value on the same request. Ignored on every method other than session/new.",
            "schema": {
              "type": "string"
            },
            "example": "work-openai"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response (success or error). Always 200 OK -- JSON-RPC errors are reported via the body's error field, not the HTTP status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header while ACPX_AUTH_TOKEN is set on the server.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/ws": {
      "get": {
        "summary": "WebSocket upgrade",
        "description": "Upgrades to a persistent, full-duplex JSON-RPC connection carrying the same Request/Response frames as POST /rpc, plus live session/update notification frames (acpx_core::notify::NotificationHub). Authorization and X-Acpx-Tenant are read once at upgrade time and apply for the connection's entire lifetime -- there is no per-message re-check after the upgrade succeeds. X-Acpx-Profile has no effect on this transport (see that parameter's own description).",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "description": "Bearer token, checked in constant time against ACPX_AUTH_TOKEN when that env var is set on the server (acpx-server/src/transport/http.rs's AuthConfig). Unset ACPX_AUTH_TOKEN means auth is disabled entirely -- this header is then ignored if sent and not required if omitted. When ACPX_AUTH_TOKEN is set, a missing or mismatched header yields 401 Unauthorized with a JSON-RPC-shaped error body (see the shared errorResponse schema on this document's responses).",
            "schema": {
              "type": "string",
              "pattern": "^Bearer .+$"
            },
            "example": "Bearer sk-acpx-..."
          },
          {
            "name": "X-Acpx-Tenant",
            "in": "header",
            "required": false,
            "description": "Self-declared tenant partition key (acpx-tenant-isolation plan) applied to every method on this request/connection, not just session/new. Absent or not valid UTF-8 is treated identically to the implicit default tenant (\"default\") -- this is a data partition, not an auth gate, so it fails open rather than rejecting the request. For WS, this header is read once at upgrade time and cached for the connection's entire lifetime (a WS client is one fixed tenant for its whole connection).",
            "schema": {
              "type": "string",
              "default": "default"
            },
            "example": "acme-corp"
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols -- upgrade succeeded."
          },
          "401": {
            "description": "Missing or invalid Authorization header while ACPX_AUTH_TOKEN is set on the server.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Request": {
        "type": "object",
        "properties": {
          "jsonrpc": {
            "$ref": "#/components/schemas/JsonRpcVersion"
          },
          "id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RequestId"
              },
              {
                "type": "null"
              }
            ]
          },
          "method": {
            "type": "string"
          },
          "params": true
        },
        "required": [
          "jsonrpc",
          "method"
        ],
        "description": "A JSON-RPC 2.0 request or notification (notifications omit `id`)."
      },
      "JsonRpcVersion": {
        "type": "string",
        "const": "2.0",
        "description": "Always the literal string \"2.0\"."
      },
      "RequestId": {
        "anyOf": [
          {
            "type": "integer",
            "format": "int64"
          },
          {
            "type": "string"
          }
        ],
        "description": "JSON-RPC ids are either a string or a number on the wire."
      },
      "Response": {
        "type": "object",
        "properties": {
          "jsonrpc": {
            "$ref": "#/components/schemas/JsonRpcVersion"
          },
          "id": {
            "$ref": "#/components/schemas/RequestId"
          },
          "result": true,
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RpcError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "jsonrpc",
          "id"
        ],
        "description": "A JSON-RPC 2.0 response: exactly one of `result`/`error` is present."
      },
      "RpcError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "data": true
        },
        "required": [
          "code",
          "message"
        ]
      },
...
    }
  }
}